215. module ui
We introduce the ui module. This is a big commit, but the code here is
"horizontal", not "vertical", so the size isn't a big deal.

Some of the concerns here:

* We can create widgets in Lua.
* We can access widgets created in C.
* Widgets have methods and properties (modules/ui.c, ui.lua).
* Garbage collecting for widgets (ui/gc.lua).

Read 'ui.c' first: it's easy to see what goes on there. Then you may want
to read 'ui.lua', which complements 'ui.c' with things that were easier
to write in a high level scripting language (like the layout manager).

For low-level stuff, see ui-impl.c and ui/gc.lua.